Refactoring on fromLLM: make names shorter
