+++ src/app/features/editor/fromLLM/chatGPT2llm/chat-gpt-2-llm.spec.ts
@@ -1,0 +1,7 @@
+import { ChatGpt2Llm } from './chat-gpt-2-llm';
+
+describe('ChatGpt2Llm', () => {
+  it('should create an instance', () => {
+    expect(new ChatGpt2Llm()).toBeTruthy();
+  });
+});
