--- src/app/features/editor/fromLLM/chatGPT2llm/chat-gpt-conv2-llm-message.spec.ts
@@ -1,7 +1,0 @@
-import { ChatGptConv2LlmMessages } from './chat-gpt-conv2-llm-messages';
-
-describe('ChatGptConv2LlmMessages', () => {
-  it('should create an instance', () => {
-    expect(new ChatGptConv2LlmMessages()).toBeTruthy();
-  });
-});
