+++ src/app/features/editor/fromLLM/chatGPT2llm/conversationpicking/conversation-pick.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import {
   ConversationPickerComponent
-} from "@app/features/editor/fromLLM/conversationpicking/conversation-picker/conversation-picker.component";
-import {ChatGptConv2LlmMessages} from "@app/features/editor/fromLLM/chatGPT2llm/chat-gpt-conv2-llm-messages";
+} from "@app/features/editor/fromLLM/chatGPT2llm/conversationpicking/conversation-picker/conversation-picker.component";
+import {ChatGpt2Llm} from "@app/features/editor/fromLLM/chatGPT2llm/chat-gpt-2-llm";
 import {MatDialog} from "@angular/material/dialog";
 import {KEMLIOService} from "@app/features/editor/services/keml-io.service";
 import {IoService} from "ngx-emfular-helper";
@@ -20,7 +20,7 @@ export class ConversationPickService {
 
   startWithAllConvs(event: Event) {
     this.ioService.loadStringFromFile(event).then( (txt: string) =>
-      this.openConversationPicker(ChatGptConv2LlmMessages.separateConvs(txt))
+      this.openConversationPicker(ChatGpt2Llm.separateConvs(txt))
     )
   }
 
@@ -31,7 +31,7 @@ export class ConversationPickService {
     )
     dialogRef.componentInstance.texts = jsons;
     dialogRef.componentInstance.chosenJson.subscribe(choice => {
-      let msgs = ChatGptConv2LlmMessages.parseConversationJSON(choice)
+      let msgs = ChatGpt2Llm.parseConversationJSON(choice)
       this.kemlIOService.convFromLlmMessages(msgs)
       dialogRef.componentInstance.chosenJson.unsubscribe()
     })
